Remove unused dependency
The auto-configuration module should define only optional dependencies
as much as possible so that the starter brings the necessary bits to
enable a particular feature.

This dependency should not be compiled scope. I've removed it to test
where it is actually used and nothing breaks. Please review and either
restore it with optional scope, adding it with compile scope in the
relevant starter.