Update README.md
The output when running `az keyvault set-policy -h` (to get the help documentation) states the following for the `--secret-permissions` argument: "Space-separated list of secret permissions to assign.  Allowed values: backup, delete, get, list, purge, recover, restore, set." 

There no longer appears to be support for `all`, so I am proposing to replace this with `get list` to support the ability to `get` and `list` the keyvault secret key / values.