Refactor UserPrincipalManager to derive Key.
resolve #472.

Update the UserPrincipalManager to select the Algorithm based of
the JWT token. Inits the JWKSet in the AADAuthenticationFilter bean
construction to avoid costly I/O.

Add UserPrincipalManagerTest, jwt, and key now that the Keyset is a param
The JWT, and keys are generated from http://kjur.github.io/jsjws/tool_jwt.html.