Allow configuration of UserGroup mapping.
Currently the Windows vs Microsoft return a different response. There
is no method of easily the configuring the AAD Starter to use
different keys within the JSON Node.

This PR adds a new nested configuration property to handle this.

This also loosens the validation of the properties for the deprecated
allowed-groups property which could be removed at a later date.